chore!: extract encoder code
Separation of concerns by moving encoding logic in new class.
